// store/questions/types.ts
export enum QuestionActionTypes {
    GET_QUESTIONNARE = "get_questions",
    CHANGE_STEP = "change_step",
    SELECT_ANSWER = "select_answer",
    UPDATE_QUESTION = "update_question",
}

export interface QuestionState {
    answers: Answer;
    step: number;
}